GET and PATCH working
